requirements.txt
numpy>=1.19
keras>=2.4
pandas
sqlalchemy>=1.4
pillow>=8.2.0
pyyaml>=6.0

pycocotools
opencv-python
pytz
ipywidgets
ipython
matplotlib
tensorflow>=2.0
tqdm

panel
bokeh
holoviews

psutil
